requirement.txt
Pillow
torch
transformers
opencv-python
numpy
matplotlib
numpy
opencv-python
Pillow
wget